+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add share button for tracks
 
 # 2025.01.1